Better naming of helper class for settings
