--- lib/dry/web/environment_data.rb
@@ -1,26 +1,0 @@
-require "dry/web/file_parser"
-
-module Dry
-  module Web
-    class EnvironmentData
-      def call(root, env)
-        files(root, env).each_with_object({}) do |file, hash|
-          hash.merge!(parser.(file))
-        end
-      end
-
-      private
-
-      def parser
-        @parser ||= FileParser.new
-      end
-
-      def files(root, env)
-        [
-          root.join(".env"),
-          root.join(".env.#{env}")
-        ].compact
-      end
-    end
-  end
-end
